Changing id type in event class from int to long Creating validation message for person and transaction

+++ src/main/java/com/homeBudget/domain/event/EventRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 
 @Repository
 public interface EventRepository extends JpaRepository<Event, Integer> {
-    Event findEventById(int id);
+    Event findEventById(Long id);
     List<Event> findEventsByUser(User user);
 }
